requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
boto3==1.28.85
botocore==1.31.85
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
Deprecated==1.2.14
fastapi==0.104.1
h11==0.14.0
idna==3.4
iso8601==1.1.0
jmespath==1.0.1
jwcrypto==1.5.0
jwt==1.3.1
multipart==0.2.4
paho-mqtt==1.6.1
passlib==1.7.4
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==4.1.0
python-multipart==0.0.6
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
SQLite3-0611==0.0.1
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
wrapt==1.16.0
